(image-cloze-association): add preview for drag items on touch screen, refactor  PD-2991
